Changed order model, request, response, repo, handler and db table

+++ internal/endpoint/http/responses/order_response.go
@@ -4,7 +4,7 @@ type OrderResponse struct {
 	ID          int    `json:"id"`
 	UserID      int    `json:"userID"`
 	SupplierID  int    `json:"supplierID"`
-	DriverID    int    `json:"driverID"`
+	DriverID    int64  `json:"driverID"`
 	StatusID    int    `json:"statusID"`
 	Note        string `json:"note"`
 	CreatedAt   string `json:"createdAt"`
